Tier 3 refactoring: split nostr-auth, extract chapters hook and API handler
- Split 1,096-line components/nostr-auth.tsx into four focused files:
  index.tsx (NostrAuth core ~220 lines), account-menu.tsx (AccountMenu +
  BunkerHealthBanner + WalletButton), login-methods.tsx (OtherSignIn +
  AmberCompletion + AmberManualPaste), muted-accounts.tsx (already split)
- Extract useChapters(url) hook to lib/chapters.ts; episode-detail-view and
  fullscreen-player now import it instead of duplicating fetch+parse logic
- Extract withErrorHandling() to lib/api-handler.ts; all five API routes now
  use it, eliminating repeated try/catch boilerplate
- Fix missing useEffect/useState imports in episode-detail-view and
  fullscreen-player that were dropped during Tier 1 chapter extraction
