+++ app/api/lightning/boostbox/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { getErrorMessage } from '@/lib/util';
+import { withErrorHandling } from '@/lib/api-handler';
 
 // Server-side proxy for the BoostBox metadata service.
 // Defaults match the public reference instance so the integration works
@@ -8,7 +8,7 @@ const BOOSTBOX_URL = process.env.BOOSTBOX_URL || 'https://tardbox.com';
 const BOOSTBOX_API_KEY = process.env.BOOSTBOX_API_KEY || 'v4v4me';
 
 export async function POST(req: Request) {
-  try {
+  return withErrorHandling(async () => {
     const payload = await req.json();
     const upstream = await fetch(`${BOOSTBOX_URL}/boost`, {
       method: 'POST',
@@ -29,10 +29,5 @@ export async function POST(req: Request) {
 
     const data = await upstream.json();
     return NextResponse.json(data);
-  } catch (e) {
-    return NextResponse.json(
-      { error: getErrorMessage(e, 'BoostBox proxy failed') },
-      { status: 500 },
-    );
-  }
+  }, 'BoostBox proxy failed');
 }
