+++ components/episode-detail-view.tsx
@@ -2,6 +2,8 @@
 import { useEffect, useState } from 'react';
 import { useApp } from '@/lib/store';
 import { fmtDuration, stripHtml } from '@/lib/format';
+import { useChapters } from '@/lib/chapters';
+import type { ChapterEntry } from '@/lib/chapters';
 import { BoltIcon } from './icons';
 import { PodcastCover } from './podcast-cover';
 import { BoostModal } from './boost-modal';
@@ -9,33 +11,8 @@ import { BoostAllModal } from './boost-all-modal';
 import { EpisodeNostrFeed } from './episode-nostr-feed';
 import type { Episode, ValueBlock } from '@/lib/types';
 
-interface ChapterEntry {
-  startTime: number;
-  title?: string;
-}
-
-function ChaptersList({ episodeId, url }: { episodeId: number; url: string }) {
-  const [chapters, setChapters] = useState<ChapterEntry[] | null>(null);
-  const [loading, setLoading] = useState(false);
-
-  useEffect(() => {
-    if (chapters !== null || loading) return;
-    setLoading(true);
-    fetch(url)
-      .then((r) => (r.ok ? r.json() : null))
-      .then((data) => {
-        const list: ChapterEntry[] = Array.isArray(data?.chapters)
-          ? data.chapters.map((c: { startTime?: unknown; title?: unknown }) => ({
-              startTime: Number(c.startTime) || 0,
-              title: typeof c.title === 'string' ? c.title : undefined,
-            }))
-          : [];
-        setChapters(list);
-      })
-      .catch(() => setChapters([]))
-      .finally(() => setLoading(false));
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [episodeId, url]);
+function ChaptersList({ url }: { url: string }) {
+  const { chapters, loading } = useChapters(url);
 
   if (loading && chapters === null) {
     return <p className="text-xs text-muted">Loading chapters…</p>;
@@ -236,7 +213,7 @@ export function EpisodeDetailView() {
 
         {/* Chapters */}
         {episode.chaptersUrl && (
-          <ChaptersList episodeId={episode.id} url={episode.chaptersUrl} />
+          <ChaptersList url={episode.chaptersUrl} />
         )}
 
         {/* Show notes */}
